+++ src/ambition/Ambition_input.cc
@@ -28,6 +28,9 @@
 #define _AMBITION_IMPLEMENTATION
 #include "OMOUSE.h"
 
+#include "Ambition_user_interface.hh"
+
+
 namespace Ambition {
 
 void calculateScroll(
@@ -47,5 +50,82 @@ void calculateScroll(
     y = 1;
   }
 }
+
+namespace Input {
+
+bool detectScroll(
+  const bool enableMouse,
+  const UserInterface::Rectangle& area,
+  const std::vector<Activation>& activations,
+  const std::map<ScrollOrientation, Action>& actions,
+  const PageSize verticalPageSize,
+  const PageSize verticalMaximum,
+  const PageSize horizontalPageSize,
+  const PageSize horizontalMaximum
+) {
+  if (!area.contains({.left = mouse.cur_x, .top = mouse.cur_y,})) {
+    return false;
+  }
+
+  int scrollHorizontal = 0;
+  int scrollVertical = 0;
+
+  if (enableMouse && mouse.get_scroll(&scrollHorizontal, &scrollVertical)) {
+  } else {
+    for (const auto& activation : activations) {
+      int& scroll
+        = activation.orientation == ScrollOrientation::Horizontal
+        ? scrollHorizontal
+        : scrollVertical;
+      auto step = 1;
+      if (activation.distance == ScrollStep::Page) {
+        step
+          = activation.orientation == ScrollOrientation::Horizontal
+          ? horizontalPageSize()
+          : verticalPageSize();
+      }
+      if (activation.distance == ScrollStep::End) {
+        step
+          = activation.orientation == ScrollOrientation::Horizontal
+          ? horizontalMaximum()
+          : verticalMaximum();
+      }
+      const auto amount
+        = (activation.direction == ScrollDirection::Forward ? 1 : -1)
+        * step;
+
+      for (const auto& _7kaaDetect : activation._7kaaButtonDetects) {
+        if (_7kaaDetect()) {
+          scroll = amount;
+        }
+      }
+      for (const auto& keyCode : activation.keyCodes) {
+        if (mouse.unique_key_code == keyCode) {
+          scroll = amount;
+        }
+      }
+    }
+  }
+
+  bool actionExecuted = false;
+  if (scrollHorizontal != 0) {
+    const auto action = actions.find(ScrollOrientation::Horizontal);
+    if (action != actions.end()) {
+      action->second(scrollHorizontal);
+      actionExecuted = true;
+    }
+  }
+  if (scrollVertical != 0) {
+    const auto action = actions.find(ScrollOrientation::Vertical);
+    if (action != actions.end()) {
+      action->second(scrollVertical);
+      actionExecuted = true;
+    }
+  }
+
+  return actionExecuted;
+}
+
+} // namespace Ambition::Input
 
 } // namespace Ambition