+++ src/ambition/version_details.cc
@@ -19,6 +19,7 @@ std::vector<VersionUpdates> versionDetails = {
       {
         .heading = _(HEADING_ADDED),
         .items = {
+          _("Added save/load game list scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
           _("Added scroll up and scroll down prompts to the What's New screen."),
           _("Added scroll hints to the What's New screen."),
         },